requirements.txt
# Requisitos del proyecto Water Delivery (raíz)
# Referencia: water_delivery/requirements.txt

setuptools>=65.5.1

Django==4.2.7
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-widget-tweaks==1.5.0

python-decouple==3.8

whitenoise==6.6.0

psycopg2-binary==2.9.9

gunicorn==21.2.0

dj-database-url==2.1.0

django-debug-toolbar==4.2.0

django-cors-headers==4.3.1
